syncpoint: add support for visualization
Add all necessary data structures to the syncpoint library to be able to
visualize call sequences. We now store all recent calls in a
CircularBuffer and keep track of calling statistics (first call, last
call, frequency, wait time). Furthermore, we use that data to generate a
DOT graph. This could for example be used to show the call graph in
webview.