syncpoint: use std::string for identifiers, API change!
Using char * for SyncPoint identifiers is cumbersome because all strings
have to be copied during initialization. This also means that we would
have a fixed length for identifier strings. In our test cases, where we
start many threads simultaneously, the compiler often optimized the
identifier strings such that two different identifier strings had the
same address, which caused the identifier saved in the set of watchers
to be changed, leading to a SyncPointAlreadyOpenedException for a new
component. This can be avoided by copying all strings in the
constructor, but using std::string avoids this problem alltogether.

Note: This changes the API because all functions such as get_syncpoint,
wait, and emit expect std::strings instead of char *.