syncpoint: always reset pending emitters after SyncBarrier was released
Until now, a SyncBarrier needed a central controller which reset the
barrier every iteration. Now, reset the barrier immediately after it has
been released, i.e. after every pending emitter has emitted the barrier.
This allows us to remove any central component which controls the
barrier, and we can define a control loop independently of the main
thread.

As an example, one could reconstruct a main loop in the following way:
 --- barrier: sensor_hook --- <---\
    |              |               |
    v              v               |
 sensor1        sensor2            |
    |              |               |
    v              v               |
 --- barrier: world_hook ---       |
            |                      |
            v                      |
     world state thread            |
            |                      |
            v                      |
 --- barrier: think_hook ---       |
     |      |     |                |
     v      v     v                |
   think1 think2 think3            |
     |      |     \                |
     \______\______\_______________/

where barrier->thread are wait() calls, and thread->barrier are emit()
calls.

Note that this also means that if a thread waits on a barrier which has
just been released, the thread will wait until it is released again.