syncpoint: add SyncPoint hierarchy
The hierarchy allows to wait for a SyncPoint which is somewhere upward
in the path, e.g. a component waiting for '/some/topic' will be
unblocked if a SyncPoint '/some/topic/sp' is emitted.
This allows to wait for one of the subordinate components. Note this
corresponds to OR: If ANY of the subordinate syncpoints is emitted, the
waiter will unblock.

In order to do so, automatically add a predecessor SyncPoint during
SyncPoint creation and also automatically register the component for
that SyncPoint. Additionally, any predecessor SyncPoint is emitted if a
SyncPoint is emitted.