aspects: add SyncPointAspect
The SyncPointAspect waits for a SyncPoint before
every loop and/or emits a SyncPoint after every loop. The SyncPoints are
created and managed by the aspect.

There are two modes of operation for the input SyncPoint: If the aspect
is initialized with WakeupType::WAIT_FOR_ONE, then the SyncPoint waits
until ONE other component emits the SyncPoint. Similarly, if the aspect
is initialized with WakeupType::WAIT_FOR_ALL, then the SyncPoint waits
for ALL other components which have registered for the SyncPoint.

Depending on the initialization, the thread may
- only wait for a SyncPoint at the beginning of the loop and not emit a
  SyncPoint after the loop,
- only emit a SyncPoint at the end of the loop but not wait for a
  SyncPoint at the beginning of the loop
- both wait at the beginning of the loop and emit (a different)
  SyncPoint at the end of the loop