aspects: make the BlockedTimingAspect a SyncPointAspect
Use syncpoints for all BlockedTiming threads. For each hook, add a start
and an end hook syncpoint. At the beginning of each hook, the main
thread emits the start syncpoint. At the end of each hook, the main
thread waits for the end syncpoint, which is emitted by all threads
registered for the hook.

This replaces the main loop with syncpoints. However, the main thread
still has to manage all hooks, similar to before.

In order to use the new main loop, each blocked timing thread currently
needs to be changed to continuous mode. Otherwise, no changes are
required.