syncpoint: do not throw on multiple register calls
We want to allow a component to register for multiple syncpoints, e.g.
'/some/sp1' and '/some/sp2'. With hierarchical syncpoints, the component
will also register for the syncpoint '/' when registering for the first
syncpoint. When registering for the second syncpoint, an exception will
be thrown as the component is already registered for '/'.
Calling register for the same syncpoint multiple times doesn't do any
damage, the exception was only a precaution and can thus be safely
removed.