aspects: BlockedTimingAspect with wait-for-wakeup and syncpoints
Adapt the BlockedTimingAspect such that it expects the thread to be in
wait-for-wakeup mode while using syncpoints. Since the thread is no
longer woken up by the main thread, add a loop listener which wakes up
the thread immediately after the loop returned. This recreated the
expected behavior: After the thread finished its loop(), the loop
listener immediately wakes up the thread, which then waits for the
syncpoint at the beginning of the loop, before it continues with the
next iteration.

Since a thread is never woken up by the main thread, we still need to
start the thread, otherwise it will never run. For now, we start the
thread in the initializer of the BlockedTimingAspect.