syncpoint: allow waiter to lock all emitters
Sometimes it is necessary to force all emitters to wait before the next
emit() call until a waiter has called wait(). As an example, a thread
watching all emitters of a given syncpoint (such as the MainThread)
needs to call wait() before the emitters call emit. Thus, add a mutex as
a stopby for all emitters.

If a component needs to lock the emitters, it can simply call
lock_until_next_wait(component_id). This way, any emitter will lock
until the specified waiter calls wait().