baseapp: add a mainloop based on syncpoints to have min/max loop time
Similar to the old mainloop, this allows to configure a minimum and a
maximum for the duration of the mainloop. If the loop time is less than
the minimum, the mainloop waits until the desired minimum is reached. If
the maximum is reached, a warning is printed.

Note that in contrast to the old mainloop, this does not remove any
threads from the loop, but only prints warnings.