aspects: use std::map to translate wakeup hooks into syncpoints
Instead of using a switch() in each method, define a static const map
that contains the translation from wakeup hooks to syncpoints and use
that map in each method. This avoids inconsistent naming (e.g. syncpoint
name is changed in one place and forgotten in the other place) and
allows us to translate back from syncpoints to hooks.