syncpoint: move dot graph generation to webview plugin
Until now, the SyncPoingManager generated the (textual) dot graph for
the syncpoint visualization. However, this lead to the cyclic dependency
syncpoint -> aspect -> syncpoint. This is because the blocked timing
aspect uses syncpoints, and for visualization, we need to get all
blocked timing hooks. Furthermore, the generation of the dot graph is
not the SyncPointManager's task anyway. Thus, move the dot graph
generation to the webview plugin where it is used.

The disadvantage of this is that we cannot generate dotgraphs outside of
the webview plugin anymore. If this is necessary, the functionality
should be moved to a separate lib/plugin.