robot-memory: if try-lock throws an exception, check if we got the mutex
Sometimes, a MongoDB exception is thrown, but we still acquired the
mutex. To detect this case, try to check the database whether we
acquired the mutex, and if we did, let the try-lock succeed. Otherwise,
or if there is another exception, return false and report an error.

Thanks to Daniel Habering for debugging and suggestions on the error
handling.